backend/requirements.txt
# Flask Core
Flask
Flask-Cors

# Database
psycopg2-binary

# Environment Variables
python-dotenv
